IDE: clear Project Inspector when active project is closed
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@34560 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ ide/projectinspector.pas
@@ -629,19 +629,20 @@ var
   FilesBranch: TTreeFilterBranch;
   Filename: String;
 begin
-  if LazProject=nil then Exit;
   FilesBranch:=FilterEdit.GetBranch(FFilesNode);
-  FilterEdit.SelectedPart:=FNextSelectedPart;
-  FilterEdit.ShowDirHierarchy:=ShowDirectoryHierarchy;
-  FilterEdit.SortData:=SortAlphabetically;
-  FilterEdit.ImageIndexDirectory:=ImageIndexDirectory;
-  // collect and sort files
-  CurFile:=LazProject.FirstPartOfProject;
-  while CurFile<>nil do begin
-    Filename:=CurFile.GetShortFilename(true);
-    if Filename<>'' then
-      FilesBranch.AddNodeData(Filename, CurFile, CurFile.Filename);
-    CurFile:=CurFile.NextPartOfProject;
+  if LazProject<>nil then begin
+    FilterEdit.SelectedPart:=FNextSelectedPart;
+    FilterEdit.ShowDirHierarchy:=ShowDirectoryHierarchy;
+    FilterEdit.SortData:=SortAlphabetically;
+    FilterEdit.ImageIndexDirectory:=ImageIndexDirectory;
+    // collect and sort files
+    CurFile:=LazProject.FirstPartOfProject;
+    while CurFile<>nil do begin
+      Filename:=CurFile.GetShortFilename(true);
+      if Filename<>'' then
+        FilesBranch.AddNodeData(Filename, CurFile, CurFile.Filename);
+      CurFile:=CurFile.NextPartOfProject;
+    end;
   end;
   FilterEdit.InvalidateFilter;            // Data is shown by FilterEdit.
 end;
@@ -652,48 +653,50 @@ var
   RequiredBranch, RemovedBranch: TTreeFilterBranch;
   NodeText, AFilename: String;
 begin
-  if LazProject=nil then exit;
-
-  // required packages
   RequiredBranch:=FilterEdit.GetBranch(DependenciesNode);
-  Dependency:=LazProject.FirstRequiredDependency;
-  while Dependency<>nil do begin
-    // Figure out the item's caption
-    NodeText:=Dependency.AsString;
-    if Dependency.DefaultFilename<>'' then begin
-      AFilename:=Dependency.MakeFilenameRelativeToOwner(Dependency.DefaultFilename);
-      if Dependency.PreferDefaultFilename then
-        NodeText:=Format(lisCEIn, [NodeText,AFilename])  // like the 'in' keyword in the uses section
-      else
-        NodeText:=Format(lisPckEditDefault, [NodeText, AFilename]);
-    end;
-    // Add the required package under the branch
-    RequiredBranch.AddNodeData(NodeText, Dependency);
-    Dependency:=Dependency.NextRequiresDependency;
-  end;
-
-  // removed required packages
-  Dependency:=LazProject.FirstRemovedDependency;
-  if Dependency<>nil then begin
-    // Create root node for removed dependencies if not done yet.
-    if RemovedDependenciesNode=nil then begin
-      RemovedDependenciesNode:=ItemsTreeView.Items.Add(DependenciesNode,
-                                              lisProjInspRemovedRequiredPackages);
-      RemovedDependenciesNode.ImageIndex:=ImageIndexRemovedRequired;
-      RemovedDependenciesNode.SelectedIndex:=RemovedDependenciesNode.ImageIndex;
-    end;
-    RemovedBranch:=FilterEdit.GetBranch(RemovedDependenciesNode);
-    // Add all removed dependencies under the branch
+  Dependency:=Nil;
+  if LazProject<>nil then begin
+    // required packages
+    Dependency:=LazProject.FirstRequiredDependency;
     while Dependency<>nil do begin
-      RemovedBranch.AddNodeData(Dependency.AsString, Dependency);
+      // Figure out the item's caption
+      NodeText:=Dependency.AsString;
+      if Dependency.DefaultFilename<>'' then begin
+        AFilename:=Dependency.MakeFilenameRelativeToOwner(Dependency.DefaultFilename);
+        if Dependency.PreferDefaultFilename then
+          NodeText:=Format(lisCEIn, [NodeText,AFilename])  // like the 'in' keyword in the uses section
+        else
+          NodeText:=Format(lisPckEditDefault, [NodeText, AFilename]);
+      end;
+      // Add the required package under the branch
+      RequiredBranch.AddNodeData(NodeText, Dependency);
       Dependency:=Dependency.NextRequiresDependency;
     end;
-  end else begin
-    // No removed dependencies -> delete the root node
-    if RemovedDependenciesNode<>nil then begin
-      FilterEdit.DeleteBranch(RemovedDependenciesNode);
-      FreeThenNil(RemovedDependenciesNode); // ItemsTreeView.Items.Delete() ?
+
+    // removed required packages
+    Dependency:=LazProject.FirstRemovedDependency;
+    if Dependency<>nil then begin
+      // Create root node for removed dependencies if not done yet.
+      if RemovedDependenciesNode=nil then begin
+        RemovedDependenciesNode:=ItemsTreeView.Items.Add(DependenciesNode,
+                                                lisProjInspRemovedRequiredPackages);
+        RemovedDependenciesNode.ImageIndex:=ImageIndexRemovedRequired;
+        RemovedDependenciesNode.SelectedIndex:=RemovedDependenciesNode.ImageIndex;
+      end;
+      RemovedBranch:=FilterEdit.GetBranch(RemovedDependenciesNode);
+      // Add all removed dependencies under the branch
+      while Dependency<>nil do begin
+        RemovedBranch.AddNodeData(Dependency.AsString, Dependency);
+        Dependency:=Dependency.NextRequiresDependency;
+      end;
     end;
+  end;
+
+  // Dependency is set to removed required packages if there is active project
+  if (Dependency=nil) and (RemovedDependenciesNode<>nil) then begin
+    // No removed dependencies -> delete the root node
+    FilterEdit.DeleteBranch(RemovedDependenciesNode);
+    FreeThenNil(RemovedDependenciesNode);
   end;
 end;
